Make Collector() work with Python 2.6
In Python 2.6, the exc_val argument to __exit__() is not always an
exception instance (see http://bugs.python.org/issue7853 for details).
Add a work-around to ensure it is.

The rest of the code bases has passing tests for Python 2.6, so add
py26 to the environment list used by tox. Fixes #10.